refactor: refactored analytics stuff in backend to its own files rather than sharing with group files
