+++ backend/routes/groups.js
@@ -13,9 +13,7 @@ const { // import controller functions for route handlers
     deleteGroup,
     createUnitGroups,
     shuffleUnitGroups,
-    moveStudent,
-    getAllGroupsAnalytics,
-    getGroupAnalytics
+    moveStudent
 } = require('../controllers/groupController');
 
 const {
@@ -42,12 +40,6 @@ router.get('/:unitCode/:year/:period', getAllGroups);
 
 // get a specific group for a specific unit
 router.get('/:unitCode/:year/:period/:groupNumber', getGroup);
-
-//get all group analytics
-router.get('/:unitCode/:year/:period/analytics', getAllGroupsAnalytics)
-
-//get a specific group analytics
-router.get('/:unitCode/:year/:period/analytics/:groupNumber', getGroupAnalytics)
 
 // create unit groups
 router.post('/:unitCode/:year/:period', createUnitGroups);
